[TRTLLM-12457][test] split _torch/speculative into hw-agnostic subdir
Move 16 hardware-agnostic spec-decode test files into a new subdirectory
unittest/_torch/speculative/hw_agnostic/. test_eagle3.py stays at the
top level because it parametrizes on DeepSeek-V3-Lite/nvfp4_moe_only,
which exercises the Blackwell NVFP4 MoE kernel path.

YAML changes:
- l0_b200.yml: unittest/_torch/speculative -> .../speculative/test_eagle3.py
  (B-tier now only runs the SM-gated file; the other 16 are covered on H100)
- l0_h100.yml: -k "eagle3" / -k "not eagle3" -> .../speculative/test_eagle3.py
  + .../speculative/hw_agnostic. The two-bucket split is preserved so the
  long-running eagle3 stage stays isolated from the rest.

CSV change:
- agg_unit_mem_df.csv: B200 row updated from unittest/_torch/speculative
  to unittest/_torch/speculative/test_eagle3.py so the case-string lookup
  still matches and parallel_factor=4 stays in effect.

Files moved are confirmed agnostic by audit (no @skip_pre_blackwell, no
isSM10x checks, no fp4/fp8/mxfp4/int4/awq/gptq parametrize values, no
SM-branched cpp kernels): test_dflash, test_draft_len_schedule,
test_draft_target, test_draft_token_prepare_for_generation,
test_draft_token_tree_sampling, test_draft_token_tree_verification,
test_dynamic_spec_decode, test_kv_cache_reuse, test_mtp, test_ngram,
test_pard, test_sa, test_save_state, test_spec_gate,
test_torch_rejection_sampling, test_user_provided.

The _torch top-level layout stays simple: only speculative/ shows up,
not a separate speculative_hw_agnostic/ sibling.
