+++ tests/integration/defs/agg_unit_mem_df.csv
@@ -124,7 +124,7 @@ unittest/auto_deploy/singlegpu/transformations,NVIDIA B200,4,
 unittest/auto_deploy/singlegpu/utils,NVIDIA B200,4,
 unittest/_torch/attention,NVIDIA B200,4,
 unittest/_torch/misc,NVIDIA B200,4,
-unittest/_torch/speculative,NVIDIA B200,4,
+unittest/_torch/speculative/test_eagle3.py,NVIDIA B200,4,
 unittest/_torch/thop/parallel,NVIDIA B200,16,
 unittest/kv_cache_manager_v2_tests/,NVIDIA B200,8,
 unittest/auto_deploy/singlegpu/compile,NVIDIA H100,4,
