+++ tests/integration/test_lists/test-db/l0_h100.yml
@@ -43,8 +43,8 @@ l0_h100:
   - unittest/_torch/modules/moe/test_moe_module.py::test_configurable_moe_single_gpu -k "CUTLASS"
   - unittest/_torch/multimodal
   - unittest/_torch/sampler
-  - unittest/_torch/speculative -k "eagle3"
-  - unittest/_torch/speculative -k "not eagle3"
+  - unittest/_torch/speculative/test_eagle3.py
+  - unittest/_torch/speculative/hw_agnostic
   - unittest/_torch/thop/parallel
   - unittest/_torch/thop/serial
   # Only key models in H100: llama/mixtral/gemma3/gpt-oss
