+++ tests/unittest/_torch/speculative/hw_agnostic/test_ngram.py
@@ -6,22 +6,19 @@ import pytest
 import torch
 
 from tensorrt_llm import LLM, SamplingParams
-from tensorrt_llm._torch.speculative.ngram import NGramDrafter, NGramPoolManager
-from tensorrt_llm.llmapi import (CudaGraphConfig, KvCacheConfig,
-                                 NGramDecodingConfig,
-                                 UserProvidedDecodingConfig)
+from tensorrt_llm.llmapi import CudaGraphConfig, KvCacheConfig, NGramDecodingConfig
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
 from utils.llm_data import llm_models_root
 
 
 # TODO: add disable_overlap_scheduler=False
 @pytest.mark.parametrize(
     "disable_overlap_scheduler,use_cuda_graph,attn_backend",
-    [[True, False, "TRTLLM"], [True, True, "TRTLLM"],
-     [True, False, "FLASHINFER"]])
-def test_llama_user_provided(disable_overlap_scheduler: bool,
-                             use_cuda_graph: bool, attn_backend: str):
+    [[True, False, "TRTLLM"], [True, True, "TRTLLM"], [True, False, "FLASHINFER"]],
+)
+@pytest.mark.high_cuda_memory
+def test_llama_ngram(disable_overlap_scheduler: bool, use_cuda_graph: bool, attn_backend: str):
     total_mem_gb = torch.cuda.get_device_properties(0).total_memory / 1e9
     if total_mem_gb < 20:
         pytest.skip("Not enough memory to load target model")
@@ -29,12 +26,11 @@ def test_llama_user_provided(disable_overlap_scheduler: bool,
     max_batch_size = 2
     max_draft_len = 4
     kv_cache_config = KvCacheConfig(enable_block_reuse=False, max_tokens=8192)
-    cuda_graph_config = CudaGraphConfig(
-        batch_sizes=[1]) if use_cuda_graph else None
+    cuda_graph_config = CudaGraphConfig(batch_sizes=[1]) if use_cuda_graph else None
 
-    llm_common_config = dict( \
-        model=llm_models_root() / "llama-3.1-model" /"Meta-Llama-3.1-8B",
-        backend='pytorch',
+    llm_common_config = dict(
+        model=llm_models_root() / "llama-3.1-model" / "Meta-Llama-3.1-8B",
+        backend="pytorch",
         attn_backend=attn_backend,
         disable_overlap_scheduler=disable_overlap_scheduler,
         cuda_graph_config=cuda_graph_config,
@@ -43,7 +39,7 @@ def test_llama_user_provided(disable_overlap_scheduler: bool,
         max_num_tokens=2048,
     )
 
-    ngram_config = NGramDecodingConfig(
+    spec_config = NGramDecodingConfig(
         max_draft_len=max_draft_len,
         max_matching_ngram_size=2,
         is_keep_all=True,
@@ -51,26 +47,11 @@ def test_llama_user_provided(disable_overlap_scheduler: bool,
         is_public_pool=True,
     )
 
-    ngram_pool_manager = NGramPoolManager(
-        spec_config=ngram_config,
-        max_num_requests=max_batch_size,
-    )
-
-    drafter = NGramDrafter(
-        spec_config=ngram_config,
-        ngram_pool_manager=ngram_pool_manager,
-    )
-
-    spec_config = UserProvidedDecodingConfig(
-        max_draft_len=max_draft_len,
-        drafter=drafter,
-    )
-
     prompts = [
         "The capital of France is",
         "The president of the United States is",
     ]
-    sampling_params = SamplingParams(max_tokens=32)
+    sampling_params = SamplingParams(max_tokens=32, ignore_eos=True)
 
     llm_spec = LLM(**llm_common_config, speculative_config=spec_config)
     results_spec = llm_spec.generate(prompts, sampling_params)
